requirements.txt
numpy~=1.26.4
matplotlib~=3.9.0
scikit-learn~=1.5.0
pillow~=10.3.0
joblib~=1.4.2
